+++ src/swell/deployment/prep_suite.py
@@ -4,7 +4,7 @@
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
 
-from tidesuites.platforms import get_tasks_per_node
+from swell.deployment.platforms import get_tasks_per_node
 
 import os
