+++ src/swell/suites/modules
@@ -11,18 +11,17 @@ module load comp/intel/19.1.0.166
 module load mpi/impi/19.1.0.166
 module load jedi/intel-impi
 
-# Python module with tide dependencies
-# ------------------------------------
+# Python module with workflow dependencies
+# ----------------------------------------
 export DHOPT=/discover/nobackup/drholdaw/opt/
 module use -a $DHOPT/modulefiles/core
 module load miniconda/3.9-cylc
 
-# Load tide modules
-# -----------------
+# Load workflow modules
+# ---------------------
 export DHJEDIOPT=/discover/nobackup/drholdaw/JediOpt/
 module use -a $DHJEDIOPT/modulefiles/core
 module load solo/develop
 module load r2d2/develop
-module load tide/work
-module load tide-suites/work
-module load tide-diagnostics/work
+module load eva/work
+module load swell/work
