+++ src/swell/tasks/get_observations.py
@@ -6,7 +6,7 @@
 
 # --------------------------------------------------------------------------------------------------
 
-from tide.task_base import taskBase
+from swell.tasks.base.task_base import taskBase
 from r2d2 import fetch
 
 class GetObservations(taskBase):
